add game statistics module
